Split off types
