+++ package.json
@@ -23,8 +23,7 @@
     "test-fdb": "npm run --workspaces test && npx prisma generate --schema tests/prisma/schema.prisma && npm run build-fdb && jest --runInBand --coverage"
   },
   "bin": {
-    "dbos-sdk": "./dist/src/dbos-runtime/cli.js",
-    "dbos-cloud": "./dist/packages/cloud-cli/cli.js"
+    "dbos-sdk": "./dist/src/dbos-runtime/cli.js"
   },
   "devDependencies": {
     "@prisma/client": "^5.3.1",
@@ -43,7 +42,8 @@
     "prisma": "^5.1.1",
     "supertest": "^6.3.3",
     "ts-jest": "^29.1.1",
-    "typeorm": "^0.3.17"
+    "typeorm": "^0.3.17",
+    "typescript": "^5.2.2"
   },
   "dependencies": {
     "@koa/bodyparser": "^5.0.0",
@@ -68,12 +68,9 @@
     "koa": "^2.14.2",
     "koa-logger": "^3.2.1",
     "lodash": "^4.17.21",
-    "openapi-types": "^12.1.3",
     "pg": "^8.11.1",
     "reflect-metadata": "^0.1.13",
     "serialize-error": "8.1.0",
-    "ts-json-schema-generator": "^1.4.0",
-    "typescript": "^5.2.2",
     "uuid": "^9.0.0",
     "winston": "^3.10.0",
     "yaml": "^2.3.1"
