+++ packages/dbos-openapi/package.json
@@ -20,6 +20,7 @@
   },
   "devDependencies": {
     "@types/jest": "^29.5.3",
+    "@types/koa": "^2.14.0",
     "@types/node": "^20.6.3",
     "@types/supertest": "^2.0.12",
     "@typescript-eslint/eslint-plugin": "^6.7.2",
@@ -32,6 +33,7 @@
     "ts-jest": "^29.1.1"
   },
   "dependencies": {
+    "@koa/bodyparser": "^5.0.0",
     "@opentelemetry/api-logs": "^0.47.0",
     "@opentelemetry/core": "^1.20.0",
     "@opentelemetry/sdk-logs": "^0.47.0",
@@ -41,6 +43,10 @@
     "fast-glob": "^3.3.1",
     "jsonwebtoken": "^9.0.2",
     "jwks-rsa": "^3.1.0",
+    "koa": "^2.14.2",
+    "openapi-types": "^12.1.3",
+    "ts-json-schema-generator": "^1.5.0",
+    "typescript": "^5.3.3",
     "winston": "^3.11.0",
     "winston-transport": "^4.6.0",
     "yaml": "^2.3.4"
