+++ src/httpServer/handlerTypes.ts
@@ -1,0 +1,15 @@
+
+//////////////////////////
+/* REGISTRATION OBJECTS */
+//////////////////////////
+
+export enum APITypes {
+  GET = "GET",
+  POST = "POST"
+}export enum ArgSources {
+  DEFAULT = "DEFAULT",
+  BODY = "BODY",
+  QUERY = "QUERY",
+  URL = "URL"
+}
+
