+++ src/index.ts
@@ -62,6 +62,9 @@ export {
 
 export {
   ArgSources,
+} from "./httpServer/handlerTypes";
+
+export {
   HandlerContext,
 
   // Endpoint Parameter Decorators
