+++ tests/httpServer/server.test.ts
@@ -16,7 +16,8 @@ import { RequestIDHeader } from "../../src/httpServer/handler";
 import { WorkflowUUIDHeader } from "../../src/httpServer/server";
 import { TestKvTable, generateDBOSTestConfig, setUpDBOSTestDb } from "../helpers";
 import request from "supertest";
-import { ArgSource, ArgSources, HandlerContext } from "../../src/httpServer/handler";
+import { ArgSource, HandlerContext } from "../../src/httpServer/handler";
+import { ArgSources } from "../../src/httpServer/handlerTypes";
 import { Authentication } from "../../src/httpServer/middleware";
 import { v1 as uuidv1, validate as uuidValidate } from "uuid";
 import { DBOSConfig } from "../../src/dbos-executor";
